AI/requirements.txt
transformers
torch
fastapi
pydantic
uvicorn
aiofiles
python-multipart
pdfplumber
Pillow
python-dotenv
torch
torchvision
torchxrayvision
numpy
sacremoses 
protobuf
